# api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Unified IP Information API",
    "version": "1.0.0",
    "description": "Fork of jason5ng32/MyIP exposing a unified, machine-consumed IP information surface. Lookup data (about an arbitrary queried IP) and egress-probe data (about the deployment's own outbound IP) are strictly separated namespaces. This document covers the endpoints implemented in the current release; planned routes land here together with their implementation, never before."
  },
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": { "type": "apiKey", "in": "header", "name": "X-API-Key" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "hint": { "type": "string" }
        }
      },
      "SourceError": {
        "type": "object",
        "required": ["source", "error", "status"],
        "properties": {
          "source": { "type": "string", "description": "Provider key, e.g. maxmind / ipapi_is / ipinfo / ippure" },
          "error": { "type": "string", "description": "Machine-readable cause: api_key_missing, timeout_after_8000ms, subject_mismatch, ..." },
          "status": { "type": "integer" }
        }
      },
      "LookupResponse": {
        "type": "object",
        "required": ["schema_version", "query", "lookup"],
        "properties": {
          "schema_version": { "type": "string", "enum": ["1.0"] },
          "query": {
            "type": "object",
            "required": ["ip", "mode", "timestamp", "duration_ms"],
            "properties": {
              "ip": { "type": "string" },
              "mode": { "type": "string", "enum": ["lookup"] },
              "timestamp": { "type": "string", "format": "date-time" },
              "duration_ms": { "type": "integer" }
            }
          },
          "lookup": {
            "type": "object",
            "required": ["geo", "network", "quality", "errors"],
            "properties": {
              "geo": {
                "type": ["object", "null"],
                "description": "MaxMind local-DB shape; null when the database is unavailable (see errors[]).",
                "properties": {
                  "source": { "type": "string", "enum": ["maxmind"] },
                  "ip": { "type": "string" },
                  "city": { "type": "string" },
                  "region": { "type": "string" },
                  "country": { "type": "string" },
                  "country_name": { "type": "string" },
                  "country_code": { "type": "string" },
                  "latitude": { "type": ["number", "string"] },
                  "longitude": { "type": ["number", "string"] },
                  "asn": { "type": "string" },
                  "org": { "type": "string" },
                  "timezone": { "type": "string" }
                }
              },
              "network": {
                "type": "object",
                "properties": {
                  "asn": { "type": "string" },
                  "org": { "type": "string" }
                }
              },
              "quality": { "$ref": "#/components/schemas/ProviderQuality" },
              "errors": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/SourceError" },
                "description": "Per-source failure list. Empty array = every source succeeded; a missing quality key always has a matching entry here."
              }
            }
          }
        }
      },
      "ProviderQuality": {
        "type": "object",
        "description": "Per-provider blocks, keyed by provider name. Absent key = that provider failed (see errors[]) or is not enabled for this route. ippure and ipgeolocation appear on the self route only — the first answers about the connection peer, the second is opt-in per deployment (V1_IPPURE_ENABLED / V1_IPGEOLOCATION_ENABLED, both off by default). The remaining key-driven sources appear on every route.",
        "properties": {
          "ipapi_is": { "$ref": "#/components/schemas/IpapiIsQuality" },
          "ipinfo": { "$ref": "#/components/schemas/IpinfoQuality" },
          "ippure": { "$ref": "#/components/schemas/IppureQuality" },
          "abuseipdb": { "$ref": "#/components/schemas/AbuseIpdbQuality" },
          "ipqs": { "$ref": "#/components/schemas/IpqsQuality" },
          "ipgeolocation": { "$ref": "#/components/schemas/IpgeolocationQuality" },
          "ip2location": { "$ref": "#/components/schemas/Ip2locationQuality" },
          "ipdata": { "$ref": "#/components/schemas/IpdataQuality" }
        }
      },
      "QualityResponse": {
        "type": "object",
        "description": "Provider verdicts without the geo block — for consumers that only need proxy/VPN/abuse signals and want neither the MaxMind payload nor its attribution requirement.",
        "required": ["schema_version", "query", "quality", "errors"],
        "properties": {
          "schema_version": { "type": "string", "enum": ["1.0"] },
          "query": {
            "type": "object",
            "required": ["ip", "mode", "timestamp", "duration_ms"],
            "properties": {
              "ip": { "type": "string" },
              "mode": { "type": "string", "enum": ["quality"] },
              "timestamp": { "type": "string", "format": "date-time" },
              "duration_ms": { "type": "integer" }
            }
          },
          "quality": { "$ref": "#/components/schemas/ProviderQuality" },
          "errors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SourceError" }
          }
        }
      },
      "IpapiIsQuality": {
        "type": "object",
        "properties": {
          "is_proxy": { "type": "boolean" },
          "is_vpn": { "type": "boolean" },
          "is_tor": { "type": "boolean" },
          "is_datacenter": { "type": "boolean" },
          "is_abuser": { "type": "boolean" },
          "raw": { "type": "object" }
        }
      },
      "IpinfoQuality": {
        "type": "object",
        "properties": {
          "privacy": {
            "type": ["object", "null"],
            "description": "null when the deployment's ipinfo plan lacks the privacy add-on.",
            "properties": {
              "proxy": { "type": "boolean" },
              "vpn": { "type": "boolean" },
              "tor": { "type": "boolean" },
              "relay": { "type": "boolean" },
              "hosting": { "type": "boolean" },
              "service": { "type": "string" }
            }
          },
          "raw": { "type": "object" }
        }
      },
...
      "ProbeResponse": {
        "type": "object",
        "required": ["schema_version", "query", "probe"],
        "properties": {
          "schema_version": { "type": "string", "enum": ["1.0"] },
          "query": {
            "type": "object",
            "required": ["ip", "mode", "timestamp", "duration_ms"],
            "properties": {
              "ip": { "type": "null", "description": "Always null: a probe answers about the deployment, not about an address a caller asked for. The egress address lives in probe.egress only, so the two can never be mistaken for each other." },
              "mode": { "type": "string", "enum": ["egress-probe"] },
              "timestamp": { "type": "string", "format": "date-time" },
              "duration_ms": { "type": "integer", "description": "Time spent serving this request, not the length of the probe run behind it." }
            }
          },
          "probe": {
            "type": "object",
            "required": ["egress"],
            "properties": {
              "egress": { "$ref": "#/components/schemas/EgressProbe" }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/api/v1/ip": {
      "get": {
        "summary": "Look up the calling visitor's own IP",
        "description": "Anonymous self lookup. The inspected address comes from Express's configured trusted-proxy chain (TRUST_PROXY; exactly one loopback hop by default), not from caller-selected headers or parameters. Results aggregate every configured provider in parallel under a shared 8s deadline; per-source failures land in lookup.errors[] without failing the request.",
        "security": [],
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "en" },
            "description": "BCP-47 tag for localized MaxMind names; normalized to the GeoLite2 supported set."
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated self-lookup payload (possibly partial — check lookup.errors).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LookupResponse" } } }
          },
          "400": {
            "description": "Client IP could not be determined or is not publicly routable.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "429": {
            "description": "Rate limit exceeded (anonymous tier).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "503": {
            "description": "Every lookup source failed for this request.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/v1/ip/{ip}": {
      "get": {
        "summary": "Look up an arbitrary IP address",
        "description": "Key-gated lookup of an address the caller supplies. Requires a valid X-API-Key; anonymous callers are limited to the self route. The address is validated before any upstream call (malformed and reserved space are refused), then the same aggregation runs as for the self lookup, over the key-driven sources. IPPure is never consulted here: it answers about the direct connection peer, so it would describe the deployment's own egress.",
        "security": [{ "ApiKeyHeader": [] }],
        "parameters": [
          {
            "name": "ip",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "IPv4 or IPv6 address. Reserved space (RFC 1918, loopback, CGNAT, link-local, documentation, ...) is rejected with 400."
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "en" },
            "description": "BCP-47 tag for localized MaxMind names; normalized to the GeoLite2 supported set."
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated lookup payload (possibly partial — check lookup.errors).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LookupResponse" } } }
          },
          "400": {
            "description": "Address is missing, malformed, or not publicly routable.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "401": {
            "description": "Missing or invalid X-API-Key.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "429": {
            "description": "Rate limit exceeded (key tier).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "503": {
            "description": "Every lookup source failed for this request.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/v1/quality/{ip}": {
      "get": {
        "summary": "Provider verdicts for an arbitrary IP address",
        "description": "The same providers, address validation and failure accounting as /api/v1/ip/{ip}, with the geo block omitted. MaxMind is never consulted, so the payload carries no geo attribution obligation. Both self-route-only sources are absent here too: IPPure answers about the direct connection peer, and ipgeolocation is opt-in per deployment.",
        "security": [{ "ApiKeyHeader": [] }],
        "parameters": [
          {
            "name": "ip",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "IPv4 or IPv6 address. Reserved space (RFC 1918, loopback, CGNAT, link-local, documentation, ...) is rejected with 400."
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated quality payload (possibly partial — check errors).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QualityResponse" } } }
          },
          "400": {
            "description": "Address is missing, malformed, or not publicly routable.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "401": {
            "description": "Missing or invalid X-API-Key.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "429": {
            "description": "Rate limit exceeded (key tier).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "503": {
            "description": "Every quality source failed for this request.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/v1/probe/egress": {
      "get": {
        "summary": "Probe this deployment's own outbound IP",
        "description": "Runs IPQuality against the address this deployment egresses from, or serves the last run from cache. Answers about the server, never about the caller — the response rides under probe.egress.* precisely so it cannot be read as a lookup result. The address is masked to its top 16 bits unless ?raw=true is passed, which this route only accepts from a caller that already presented a valid key. A probe takes 30-90 seconds and contacts several hundred third-party endpoints, so results are cached (default 10 minutes) and the route carries an additional hourly cap on top of the per-minute limiter; a cached read past its TTL is served immediately with stale: true while a refresh runs behind it.",
        "security": [{ "ApiKeyHeader": [] }],
        "parameters": [
          {
            "name": "raw",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean", "default": false },
            "description": "Serve the egress address in full instead of masked. Only the exact string 'true' enables it."
          }
        ],
        "responses": {
          "200": {
            "description": "The egress report. Check probe.egress.stale to tell a fresh run from a cached one.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProbeResponse" } } }
          },
          "401": {
            "description": "Missing or invalid X-API-Key.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "429": {
            "description": "Rate limit exceeded — the per-minute key tier, or the hourly probe cap.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "502": {
            "description": "The sidecar was reached but the run failed. error names which way (probe_timeout / probe_no_json / probe_failed_to_start), and detail carries the tail of the probe's own output.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "503": {
            "description": "No sidecar is configured for this deployment (IPQUALITY_SIDECAR_URL unset).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
  }
}